Update to the latest Phalcon

+++ app/common/library/Info.php
@@ -83,7 +83,7 @@ class Info
 
     public static function uptime()
     {
-        if ($uptime = Info::proc('uptime')) {
+        if ($uptime = (int) Info::proc('uptime')) {
             $days = floor($uptime / 60 / 60 / 24);
             $hours = $uptime / 60 / 60 % 24;
             $mins = $uptime / 60 % 60;
